refactor: move single source of truth to apiclaw/scripts/apiclaw.py
- Remove shared/ directory, use apiclaw/scripts/apiclaw.py as source
- Update sync-scripts.sh, pre-commit hook, and CI to point to new path
- Sync all amazon-* skill copies from new source
